[19] TODOを削除するAPIのHandler を実装してみよう

+++ service/todo.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"strings"
 
 	"github.com/TechBowl-japan/go-stations/model"
 )
@@ -121,7 +122,26 @@ func (s *TODOService) UpdateTODO(ctx context.Context, id int64, subject, descrip
 
 // DeleteTODO deletes TODOs on DB by ids.
 func (s *TODOService) DeleteTODO(ctx context.Context, ids []int64) error {
+	if len(ids) == 0 {
+		return fmt.Errorf("ids is empty")
+	}
 	const deleteFmt = `DELETE FROM todos WHERE id IN (?%s)`
+	query := fmt.Sprintf(deleteFmt, strings.Repeat(",?", len(ids)-1))
+	args := make([]interface{}, len(ids))
+	for i, id := range ids {
+		args[i] = id
+	}
+	result, err := s.db.ExecContext(ctx, query, args...)
+	if err != nil {
+		return err
+	}
+	rows, err := result.RowsAffected()
+	if err != nil {
+		return err
+	}
+	if rows == 0 {
+		return &model.ErrNotFound{Message: "TODO not found"}
+	}
 
 	return nil
 }
